add lateral join support
implementing lateral join support requires two separate but
complementary pieces. A contrived example of a lateral join in sql looks something like:

"SELECT t1.a, t2.b
FROM A a INNER JOIN LATERAL (SELECT b from B b) t2 ON (t1.a = t2.b)"

To replicate this syntax the first step is to encode the join type (inner, outer, left, right).
This is done by embedding the join type into the `Subquery`
node, which, when the subquery is built, prefixes the generated subquery
statement with, for example, "LEFT OUTER JOIN LATERAL (...)".

The second piece is capturing the `ON` clause conditions. Initial
approach was to mimic existing standard joins, but this can't work since
the subquery is not yet materialized, so `ON` clause references refer to
table refs *within* the subquery, whereas we need to reference outer
scope.

The solution is admittedly a hack, but in the scheme of things (namely
this library) it fits perfectly. Each `Subquery` node encodes its parent
query; we can pass this as the reference (`RefId`) to a new
`QueryModifier` type named `Lateral`, along with the `ON` clause
conditions (which come *after* the suquery phase in `Query`
flatMap chain, so conditions refer to outer scope).

Finally, when the subquery statement is built we check
existing `QueryModifier`s for a `Lateral` who's reference matches that
of the query embedded in `Subquery` node; if there's a match, grab the
`ON` clause node and display it.

The commits show it all, but this provides the rationale for the
implementation. Neither Quill nor Slick support lateral joins, and now I
see why -- non-trivial to implement...